refactor: move `CheckedCastType` from `Ast` to `shared`
