+++ main/src/ca/uwaterloo/flix/language/phase/TypeReconstruction.scala
@@ -15,8 +15,8 @@
  */
 package ca.uwaterloo.flix.language.phase
 
-import ca.uwaterloo.flix.language.ast.Ast.CheckedCastType
 import ca.uwaterloo.flix.language.ast.Type.getFlixType
+import ca.uwaterloo.flix.language.ast.shared.CheckedCastType
 import ca.uwaterloo.flix.language.ast.{Ast, KindedAst, Type, TypeConstructor, TypedAst}
 import ca.uwaterloo.flix.language.phase.unification.Substitution
 import ca.uwaterloo.flix.language.errors.TypeError
